requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
dimod==0.11.1
diskcache==5.4.0
dwave-cloud-client==0.10.1
dwave-greedy==0.2.5
dwave-hybrid==0.6.7
dwave-inspector==0.2.11
dwave-neal==0.5.9
dwave-networkx==0.8.12
dwave-ocean-sdk==5.1.0
dwave-preprocessing==0.4.0
dwave-system==1.15.0
dwave-tabu==0.4.5
dwavebinarycsp==0.2.0
fasteners==0.17.3
Flask==2.1.2
homebase==1.0.1
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
minorminer==0.2.8.post1
networkx==2.8.3
numpy==1.22.4
penaltymodel==1.0.2
plucky==0.4.3
pydantic==1.9.1
PySocks==1.7.1
python-dateutil==2.8.2
rectangle-packer==2.0.1
requests==2.27.1
scipy==1.8.1
six==1.16.0
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
